[wip] Create a LifeCycleCtx from an UpdateCtx
This lets container widgets create a LifeCycleCtx from within
an update call, so that they can manually send LifeCycle events
when needed.

Is this the right move? Or do we want this to be more automatic?

this requires adding `command_queue` to UpdateCtx, although
we don't really need it there?